moving meta into its own module
